feat(ui): Present error within amount view controller instead of alert
