+++ DashWallet/Sources/Categories/UIViewController+AlertPresenting.swift
@@ -1,0 +1,29 @@
+//
+//  Created by tkhp
+//  Copyright © 2022 Dash Core Group. All rights reserved.
+//
+//  Licensed under the MIT License (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  https://opensource.org/licenses/MIT
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import UIKit
+
+extension UIViewController {
+    func showAlert(with title: String?, message: String?, presentingViewController: UIViewController? = nil) {
+        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
+        let okAction = UIAlertAction(title: NSLocalizedString("OK", comment: ""), style: .cancel)
+        alert.addAction(okAction)
+
+        let presentingViewController = presentingViewController ?? self
+        presentingViewController.present(alert, animated: true)
+    }
+}
